requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.5
pandas==1.5.2
pkcs7==0.1.2
pyarrow==10.0.1
pycryptodome==3.16.0
PyP100==0.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
